group all methods that create files for a material into material_files.py
